Fix test_b010 parameter count
As stated by the scmi testlist documentation, test_b010 is testing for
an invalid command by first issuing it, and then by verifying the
return status code from the PROTOCOL_MESSAGE_ATTRIBUTES command.

https://github.com/ARM-software/scmi-tests/blob/master/docs/scmi_testlist.md

The PROTOCOL_MESSAGE_ATTRIBUTES requires the message_id to be forwarded
as an argument, hence the num_parameter in the val_send_message should
be set to 1 to make sure the message is constructed with the appropriate
length.

Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>